requirements.txt
requests==2.26.0
six==1.16.0
PyChef==0.3.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
ConfigArgParse==1.5.3
cryptography==35.0.0
idna==3.2
paramiko==2.7.2
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
sshtunnel==0.4.0
urllib3==1.26.7
